pyproject.toml
[tool.poetry]
name = "slack-bot-arxiv"
version = "0.1.0"
description = ""
authors = ["ShotaTotsuka <shota.totsuka@blueish.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
slack-bolt = "^1.19.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
